+++ CHANGES.md
@@ -3,6 +3,7 @@
 ## ?
 
 - fix: position fixes for context menu
+- fix: disable context menu more actively
 - fix: ensure context menu visible
 - fix: build issues
 - refactor: use ESM in main